# schemas/Login.go
package schemas

type Login struct {
    CustomerID string `example:"5dd1f36f-1627-4c88-98fb-601feb9634be"`
}

type LoginResponse struct {
    ID         string `example:"9cc1g36t-6537-4d10-11fs-771may9224vv"`
    CustomerID string `example:"5dd1f36f-1627-4c88-98fb-601feb9634be"`
}